+++ main.py
@@ -3,8 +3,7 @@ import re
 import discord
 from discord.ext import commands
 from datetime import datetime, timedelta
-# import params_rs as params
-from experiment import params as params
+from params_rs import params
 from redstar import Rs
 from keep_awake import keep_awake # used to keep the server awake otherwise it goes to sleep after 1h of inactivity
 import dotenv
