+++ params_rs.py
